requirements.txt
flask
flask_sqlalchemy
flask_login
sqlalchemy
pymysql
sentence-transformers
pdfminer.six
python-docx
docx2txt
requests
werkzeug
numpy
